Split state_cache service.
